Stubbed out repl console
Code is in C as the esp32 console api
isn't designed to easily wrapp C++ code around it.

The wifi task is stubbed to validate we can join a network.
Will be rework to handle more config.

Added a task command to display the running tasks stats.
Lifted from an esp32 example.